option: add options handling
This commit adds options handling for the input user arguments.
With that, the checks for file, active timeout, inactive timeout
and entries number were implemented. The util functions for converting
a string into numeric value in unsigned int data type and checking
the value is in the range are included.