netflow: add creating and saving flows
This commit adds creating flows and saving them into the binary
search tree. This version is without possible memory leaks and checked
that number of packets and flows processed are valid.
The number of flows in a tree is the same so the tree is validly filled
with created flows.