netflow: fix sent flow record structure
This commit changes the usage of the structure netflow_v5_flow_record.
Previously it was used as a tree node record as well. Anyway,
due to problems with sending flows because of invalid size
with struct timeval for first and last, the structure was duplicated
and modified to the new structure flow_node, which is used
in the flow tree.