netflow: fix exporting all by the time
This commit fixes exporting all flows at the end of the program
in the order by time from the oldest one. At the same time, the function
bst_find_oldest function was implemented which will be used for the -m
program option as well. TODO: exporting expired nodes by time
(because of active timer or inactive timer or TCP flags).