tree: add function for moving node
This commit adds a separated function from moving a node from one tree
to a new one with the deletion from the original tree. This function
is used for adding the flow to the tree of expired flows and removing
the node from the flow records tree. The usage of the function is added
for the active timer too.