netflow: change m parameter meaning
This commit changes the meaning of the m parameter. Instead
of exporting the oldest flow while the cached entries number equals
the parameter value, in the new version, the oldest flow is exported
when theoretically the number of cached flows exceeds the parameter
value by one. At the same time, all fixes based on the e-learning forum
questions are done.